added rule to stop “VCS” looking like “Vcs” in forms
Acronym inflections are not supported in this version of Rails – they were added in https://github.com/rails/rails/pull/2029 – so this will not work unless Rails is upgraded.

I haven’t tested this.

Another option is to explicitly write “VCS” in the `f.label` calls in app/views/projects/_form.html.haml . That is less DRY but would work without upgrading Rails.